+++ spec/Pim/Bundle/TransformBundle/Normalizer/Flat/ProductValueNormalizerSpec.php
@@ -2,17 +2,26 @@
 
 namespace spec\Pim\Bundle\TransformBundle\Normalizer\Flat;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use PhpSpec\ObjectBehavior;
+use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Pim\Bundle\CatalogBundle\Model\Association;
+use Pim\Bundle\CatalogBundle\Entity\Family;
+use Pim\Bundle\CatalogBundle\Entity\Group;
+use Pim\Bundle\CatalogBundle\Entity\AssociationType;
+use Pim\Bundle\CatalogBundle\Entity\Attribute;
+use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
+use Pim\Bundle\TransformBundle\Normalizer\Filter\NormalizerFilterInterface;
+
+use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Serializer\Exception\InvalidArgumentException;
+use Doctrine\Common\Collections\ArrayCollection;
 
 class ProductValueNormalizerSpec extends ObjectBehavior
 {
-    function let(SerializerInterface $serializer, AttributeInterface $simpleAttribute)
-    {
+    function let(SerializerInterface $serializer, AttributeInterface $simpleAttribute) {
         $serializer->implement('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
         $this->setSerializer($serializer);
 
@@ -42,38 +51,43 @@ class ProductValueNormalizerSpec extends ObjectBehavior
         $this->supportsNormalization(1, 'csv')->shouldBe(false);
     }
 
-    function it_normalizes_a_value_with_null_data(ProductValueInterface $value, $simpleAttribute)
+    function it_normalizes_a_value_with_null_data(ProductValueInterface $value, AttributeInterface $simpleAttribute)
     {
         $value->getData()->willReturn(null);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => '']);
     }
 
-    function it_normalizes_a_value_with_a_integer_data(ProductValueInterface $value, $simpleAttribute)
+    function it_normalizes_a_value_with_a_integer_data(ProductValueInterface $value, AttributeInterface $simpleAttribute)
     {
         $value->getData()->willReturn(12);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => '12']);
     }
 
-    function it_normalizes_a_value_with_a_float_data(ProductValueInterface $value, $simpleAttribute)
+    function it_normalizes_a_value_with_a_float_data(ProductValueInterface $value, AttributeInterface $simpleAttribute)
     {
         $value->getData()->willReturn(12.25);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => '12.2500']);
     }
 
-    function it_normalizes_a_value_with_a_string_data(ProductValueInterface $value, $simpleAttribute)
+    function it_normalizes_a_value_with_a_string_data(ProductValueInterface $value, AttributeInterface $simpleAttribute)
     {
         $value->getData()->willReturn('my data');
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => 'my data']);
     }
 
-    function it_normalizes_a_value_with_a_boolean_data(ProductValueInterface $value, $simpleAttribute)
+    function it_normalizes_a_value_with_a_boolean_data(ProductValueInterface $value, AttributeInterface $simpleAttribute)
     {
         $value->getData()->willReturn(false);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => '0']);
 
         $value->getData()->willReturn(true);
@@ -81,35 +95,29 @@ class ProductValueNormalizerSpec extends ObjectBehavior
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => '1']);
     }
 
-    function it_normalizes_a_value_with_a_collection_data(ProductValueInterface $value, $simpleAttribute, $serializer)
+    function it_normalizes_a_value_with_a_collection_data(ProductValueInterface $value, AttributeInterface $simpleAttribute, SerializerInterface $serializer)
     {
         $itemOne = new \stdClass();
         $itemTwo = new \stdClass();
         $collection = new ArrayCollection([$itemOne, $itemTwo]);
         $value->getData()->willReturn($collection);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
 
-        $serializer
-            ->normalize($collection, 'flat', ['field_name' => 'simple'])
-            ->shouldBeCalled()
-            ->willReturn(['simple' => 'red, blue']);
-
+        $serializer->normalize($collection, 'flat', ['field_name' => 'simple'])->shouldBeCalled()->willReturn(['simple' => 'red, blue']);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => 'red, blue']);
     }
 
-    function it_normalizes_a_value_with_an_array_data(ProductValueInterface $value, $simpleAttribute, $serializer)
+    function it_normalizes_a_value_with_an_array_data(ProductValueInterface $value, AttributeInterface $simpleAttribute, SerializerInterface $serializer)
     {
         $itemOne = new \stdClass();
         $itemTwo = new \stdClass();
         $array = [$itemOne, $itemTwo];
         $value->getData()->willReturn($array);
         $value->getAttribute()->willReturn($simpleAttribute);
+        $simpleAttribute->isLocaleSpecific()->willReturn(false);
 
-        $serializer
-            ->normalize(Argument::any(), 'flat', ['field_name' => 'simple'])
-            ->shouldBeCalled()
-            ->willReturn(['simple' => 'red, blue']);
-
+        $serializer->normalize(Argument::any(), 'flat', ['field_name' => 'simple'])->shouldBeCalled()->willReturn(['simple' => 'red, blue']);
         $this->normalize($value, 'flat', [])->shouldReturn(['simple' => 'red, blue']);
     }
 }