+++ src/Pim/Bundle/TransformBundle/Builder/FieldNameBuilder.php
@@ -23,16 +23,31 @@ class FieldNameBuilder
     /** @var string */
     protected $attributeClass;
 
+    /** @var string */
+    protected $channelClass;
+
+    /** @var string */
+    protected $localeClass;
+
     /**
      * @param ManagerRegistry $managerRegistry
      * @param string          $assocTypeClass
      * @param string          $attributeClass
+     * @param string          $channelClass
+     * @param string          $localeClass
      */
-    public function __construct(ManagerRegistry $managerRegistry, $assocTypeClass, $attributeClass)
-    {
+    public function __construct(
+        ManagerRegistry $managerRegistry,
+        $assocTypeClass,
+        $attributeClass,
+        $channelClass,
+        $localeClass
+    ) {
         $this->managerRegistry = $managerRegistry;
         $this->assocTypeClass  = $assocTypeClass;
         $this->attributeClass  = $attributeClass;
+        $this->channelClass    = $channelClass;
+        $this->localeClass     = $localeClass;
     }
 
     /**
@@ -45,15 +60,15 @@ class FieldNameBuilder
         $fieldNames = [];
         $assocTypes = $this->getRepository($this->assocTypeClass)->findAll();
         foreach ($assocTypes as $assocType) {
-            $fieldNames[] = $assocType->getCode() .'-groups';
-            $fieldNames[] = $assocType->getCode() .'-products';
+            $fieldNames[] = $assocType->getCode().'-groups';
+            $fieldNames[] = $assocType->getCode().'-products';
         }
 
         return $fieldNames;
     }
 
     /**
-     * Extract attribute field name informations with attribute code, locale code, scope code
+     * Extract attribute field name information with attribute code, locale code, scope code
      * and optionally price currency
      *
      * Returned array like:
@@ -78,15 +93,17 @@ class FieldNameBuilder
 
         if (null !== $attribute) {
             $this->checkFieldNameTokens($attribute, $fieldName, $explodedFieldName);
+            $attributeInfos = $this->extractAttributeInfos($attribute, $explodedFieldName);
+            $this->checkFieldNameLocaleByChannel($attribute, $fieldName, $attributeInfos);
 
-            return $this->extractAttributeInfos($attribute, $explodedFieldName);
+            return $attributeInfos;
         }
 
         return null;
     }
 
     /**
-     * Extract informations from an attribute and exploded field name
+     * Extract information from an attribute and exploded field name
      * This method is used from extractAttributeFieldNameInfos and can be redefine to add new rules
      *
      * @param AttributeInterface $attribute
@@ -112,7 +129,7 @@ class FieldNameBuilder
     }
 
     /**
-     * Extract field name informations from a potential association field name
+     * Extract field name information from a potential association field name
      *
      * Returned array like:
      * [
@@ -145,12 +162,11 @@ class FieldNameBuilder
     protected function checkFieldNameTokens(AttributeInterface $attribute, $fieldName, array $explodedFieldName)
     {
         // the expected number of tokens in a field may vary,
-        //  - with the current price import, the currency can be optionaly present in the header,
+        //  - with the current price import, the currency can be optionally present in the header,
         //  - with the current metric import, a "-unit" field can be added in the header,
         //
         // To avoid BC break, we keep the support in this fix, a next minor version could contain only the
         // support of currency code in the header and metric in a single field
-        $expectedSize = [0];
         $isLocalizable = $attribute->isLocalizable();
         $isScopable = $attribute->isScopable();
         $isPrice = 'prices' === $attribute->getBackendType();
@@ -192,12 +208,47 @@ class FieldNameBuilder
 
             throw new \InvalidArgumentException(
                 sprintf(
-                    'The field "%s" is not well-formated, attribute "%s" expects %s',
+                    'The field "%s" is not well-formatted, attribute "%s" expects %s',
                     $fieldName,
                     $attribute->getCode(),
                     $expected
                 )
             );
+        }
+        if ($isLocalizable) {
+            $this->checkForLocaleSpecificValue($attribute, $explodedFieldName);
+        }
+    }
+
+    /**
+     * Check the consistency of the field with channel associated
+     *
+     * @param AttributeInterface $attribute
+     * @param string             $fieldName
+     * @param array              $attributeInfos
+     *
+     * @throws \InvalidArgumentException
+     */
+    protected function checkFieldNameLocaleByChannel(AttributeInterface $attribute, $fieldName, array $attributeInfos)
+    {
+        if ($attribute->isScopable() &&
+            $attribute->isLocalizable() &&
+            isset($attributeInfos['scope_code']) &&
+            isset($attributeInfos['locale_code'])
+        ) {
+            $channel = $this->getRepository($this->channelClass)->findByReference($attributeInfos['scope_code']);
+            $locale = $this->getRepository($this->localeClass)->findByReference($attributeInfos['locale_code']);
+
+            if ($channel !== null && $locale !== null && !$channel->hasLocale($locale)) {
+                throw new \InvalidArgumentException(
+                    sprintf(
+                        'The locale "%s" of the field "%s" is not available in scope "%s"',
+                        $attributeInfos['locale_code'],
+                        $fieldName,
+                        $attributeInfos['scope_code']
+                    )
+                );
+            }
         }
     }
 
@@ -209,5 +260,28 @@ class FieldNameBuilder
     protected function getRepository($entityClass)
     {
         return $this->managerRegistry->getRepository($entityClass);
+    }
+
+    /**
+     * Check if provided locales for an locale specific attribute exist
+     *
+     * @param AttributeInterface $attribute
+     * @param array              $explodedFieldNames
+     */
+    protected function checkForLocaleSpecificValue(AttributeInterface $attribute, array $explodedFieldNames)
+    {
+        if ($attribute->isLocaleSpecific()) {
+            $attributeInfo = $this->extractAttributeInfos($attribute, $explodedFieldNames);
+            $availableLocales = $attribute->getLocaleSpecificCodes();
+            if (!in_array($explodedFieldNames[1], $availableLocales)) {
+                throw new \LogicException(
+                    sprintf(
+                        'The provided specific locale "%s" does not exist for "%s" attribute ',
+                        $attributeInfo['locale_code'],
+                        $attribute->getCode()
+                    )
+                );
+            }
+        }
     }
 }